+++ source/Engine/Vector.h
@@ -7,16 +7,16 @@ struct Vector{
 
     Vector operator+(const Vector& a);
     Vector operator-(const Vector& a);
-    Vector operator*(const int& a);
-    Vector operator/(const int& a);
+    Vector operator*(const float& a);
+    Vector operator/(const float& a);
 
     bool operator==(const Vector& a);
 
     // Assignment operators
     void operator+=(const Vector& a);
     void operator-=(const Vector& a);
-    void operator*=(const int& a);
-    void operator/=(const int& a);
+    void operator*=(const float& a);
+    void operator/=(const float& a);
 
     float x;
     float y;
